fix(updater): keep install stages out of the store's import graph
The macOS release dry run caught this: AutoUpdaterService.test.ts mocked
store.js with an importOriginal() factory to avoid duplicating the stage list,
but importOriginal loads the REAL store.ts, which eagerly imports electron.
On the release workflow's unit-test job the Electron binary isn't installed,
so getElectronPath() threw and the whole suite failed to load. It passed
locally only because a local Electron install is present.

Move PENDING_UPDATE_INSTALL_STAGES into electron/utils/updateInstallStages.ts,
a module with zero imports. store.ts takes the type, AutoUpdaterService takes
the constant, and the test imports the same single source of truth directly —
so the mock is a plain factory again and nothing has to reproduce the list.

Refs #11481
