+++ electron/services/AutoUpdaterService.ts
@@ -10,8 +10,9 @@ import { broadcastToRenderer } from "../ipc/utils.js";
 import { requestGracefulShutdownForUpdate } from "../lifecycle/shutdownCoordinator.js";
 import { getSystemSleepService } from "./SystemSleepService.js";
 import { trackEvent } from "./TelemetryService.js";
-import { store, PENDING_UPDATE_INSTALL_STAGES } from "../store.js";
-import type { PendingUpdateInstallStage } from "../store.js";
+import { store } from "../store.js";
+import { PENDING_UPDATE_INSTALL_STAGES } from "../utils/updateInstallStages.js";
+import type { PendingUpdateInstallStage } from "../utils/updateInstallStages.js";
 import { PRODUCT_NAME, PRODUCT_WEBSITE } from "../utils/productBranding.js";
 import { isTrustedRendererUrl } from "../../shared/utils/trustedRenderer.js";
 import { isWindowsStoreBuild } from "../../shared/config/distribution.js";
